Create README.md
Learning micropython by creating a project. Trying to track the azimuth and elevation in realtime of the sun from a fixed place on earth and displaying it on an oled screen as well as using two servos to track azimuth and elevation. I am using a Pi Pico, Ds3231 RTC, and servo motors.